+++ Assets/_Scenes/TestScenes/Julian/Puzzles/FroggerPuzzle/BuildingBlocks/Bytestream.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &5582896792403178432
+--- !u!1 &8563802780868391526
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,61 +8,59 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 5582896792403178435}
-  - component: {fileID: 5582896792403178434}
-  - component: {fileID: 5582896792403178437}
-  - component: {fileID: 5323834423315797836}
-  m_Layer: 12
-  m_Name: MovingWall
+  - component: {fileID: 9052581746606576321}
+  - component: {fileID: 2854635397422262187}
+  - component: {fileID: 7805011728143187768}
+  - component: {fileID: 6909330894814257293}
+  m_Layer: 14
+  m_Name: Bytestream
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &5582896792403178435
+--- !u!4 &9052581746606576321
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5582896792403178432}
-  m_LocalRotation: {x: 0, y: 0, z: 0.7071068, w: 0.7071068}
-  m_LocalPosition: {x: 1.93, y: 0.42, z: 0}
-  m_LocalScale: {x: 0.2627721, y: 1.9677879, z: 0.53154}
+  m_GameObject: {fileID: 8563802780868391526}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: -1
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 90}
---- !u!114 &5582896792403178434
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2854635397422262187
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5582896792403178432}
+  m_GameObject: {fileID: 8563802780868391526}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 743bc41e36b984f4cb92cac188ca7d1c, type: 3}
+  m_Script: {fileID: 11500000, guid: 8a3e7b1e18b8f9546b343018d3a1effe, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  collisionCollider: {fileID: 5582896792403178437}
-  movementMode: 1
-  waitTime: 0.64
-  easeAmount: 0.9
-  speed: 5.26
-  localWaypoints:
-  - position: {x: 0, y: 0}
-    angle: 0
-  - position: {x: -4, y: 0}
-    angle: 0
---- !u!61 &5582896792403178437
-BoxCollider2D:
+  _Collider: {fileID: 7805011728143187768}
+  _SpriteRenderer: {fileID: 6909330894814257293}
+  _ObstacleLayer:
+    serializedVersion: 2
+    m_Bits: 12288
+  _MoveSpeed: 3
+  _MoveDirection: 2
+  _Size: {x: 1, y: 3}
+--- !u!70 &7805011728143187768
+CapsuleCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5582896792403178432}
+  m_GameObject: {fileID: 8563802780868391526}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
@@ -70,25 +68,15 @@ BoxCollider2D:
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1, y: 1}
-    newSize: {x: 1, y: 1}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 1, y: 1}
-  m_EdgeRadius: 0
---- !u!212 &5323834423315797836
+  m_Size: {x: 1, y: 3}
+  m_Direction: 0
+--- !u!212 &6909330894814257293
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5582896792403178432}
+  m_GameObject: {fileID: 8563802780868391526}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -120,15 +108,15 @@ SpriteRenderer:
   m_AutoUVMaxDistance: 0.5
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: -346827107
-  m_SortingLayer: -1
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 7482667652216324306, guid: 311925a002f4447b3a28927169b83ea6, type: 3}
-  m_Color: {r: 0.4716981, g: 0.4716981, b: 0.4716981, a: 1}
+  m_Sprite: {fileID: 21300000, guid: 929954561fb02ff47be23b8c6c0f409c, type: 3}
+  m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
-  m_DrawMode: 0
-  m_Size: {x: 1, y: 1}
+  m_DrawMode: 1
+  m_Size: {x: 1, y: 3}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
