--- Assets/_Scenes/TestScenes/Julian/Puzzles/FroggerPuzzle/BuildingBlocks/TeleportationPair.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8ff8f80edac742c44b8f0a967791902a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
