+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/Checkpoint.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Puzzles
+{
+    public class Checkpoint : MonoBehaviour, IPuzzleInteractable
+    {
+    }
+}
