--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/CheckpointObstacle.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-namespace Puzzles
-{
-    public class CheckpointObstacle : MonoBehaviour, IPuzzleObstacle
-    {
-    }
-}
