--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/FroggerGoal.cs
@@ -1,5 +1,0 @@
-using UnityEngine;
-
-public class FroggerGoal : MonoBehaviour, IPuzzleObstacle
-{
-}
