--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/FroggerLane.cs
@@ -1,6 +1,0 @@
-﻿using UnityEngine;
-
-public class FroggerLane : MonoBehaviour
-{
-    
-}
