--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/FroggerLane.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2eba60bdd15d49c3b3b2b9afcae6fc76
-timeCreated: 1673222996
