--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/ImpassableObstacle.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ImpassableObstacle : MonoBehaviour, IPuzzleObstacle
-{
-}
