+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/ImpassableWall.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Puzzles
+{
+    public class ImpassableWall : MonoBehaviour, IPuzzleInteractable
+    {
+    }
+}
