--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Frogger/TeleporterObstacle.cs
@@ -1,21 +1,0 @@
-using UnityEngine;
-
-public class TeleporterObstacle : MonoBehaviour, IPuzzleObstacle
-{
-    public Transform  _Exit;
-    public Collider2D _Collider;
-    
-    public void Warp(Transform target)
-    {
-        target.position = _Exit.transform.position;
-    }
-
-
-    private void OnDrawGizmos()
-    {
-        if (_Exit != null && _Collider != null)
-        {
-            Gizmos.DrawWireSphere(_Exit.position, _Collider.bounds.size.x / 2.0f);
-        }
-    }
-}
