+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/IPuzzleInteractable.cs
@@ -1,0 +1,8 @@
+
+
+namespace Puzzles
+{
+    public interface IPuzzleInteractable
+    {
+    }
+}
