--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/IPuzzleObstacle.cs
@@ -1,4 +1,0 @@
-
-public interface IPuzzleObstacle 
-{
-}
