--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Path/PathPuzzleGoal.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PathPuzzleGoal : MonoBehaviour, IPuzzleObstacle
-{
-}
