--- Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/Path/PathPuzzleGoal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d47ac77e5efa5844f808582a58c6717c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
