+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/PuzzleBase.cs
@@ -1,18 +1,19 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-public abstract class PuzzleBase : MonoBehaviour
+namespace Puzzles
 {
-    public abstract void Init();
-    
-    public Canvas WindowCanvas { get; set; }
-    
-    public event Action<bool> onPuzzleComplete;
-
-    protected void TriggerPuzzleComplete(bool wasSuccess)
+    public abstract class PuzzleBase : MonoBehaviour
     {
-        onPuzzleComplete?.Invoke(wasSuccess);
+        public abstract void Init();
+
+        public Canvas WindowCanvas { get; set; }
+
+        public event Action<bool> onPuzzleComplete;
+
+        protected void TriggerPuzzleComplete(bool wasSuccess)
+        {
+            onPuzzleComplete?.Invoke(wasSuccess);
+        }
     }
 }
