+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/PuzzleGoal.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Puzzles
+{
+    public class PuzzleGoal : MonoBehaviour, IPuzzleInteractable
+    {
+    }
+}
