+++ Assets/_Scenes/TestScenes/Julian/Scripts/Puzzles/PuzzleWindow.cs
@@ -2,182 +2,192 @@ using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 
-public class PuzzleWindow : MonoBehaviour
+namespace Puzzles
 {
+    public class PuzzleWindow : MonoBehaviour
+    {
 #region Editor Properties
-    [Header("Scriptable Objects")]
-    public ScriptableEventDispatcher _GameEventDispatcher;
-    public PuzzleDictionary    _PuzzleDictionary;
 
-    [Header("Asset References")]
-    public GameObject      _BaseView;
-    public TextMeshProUGUI _AttemptsLeftText;
-    public TextMeshProUGUI _LevelsText;
-    
-    [Header("Tuning Parameters")]
-    public int _MaxSolveAttempts;
+        [Header("Scriptable Objects")]
+        public ScriptableEventDispatcher _GameEventDispatcher;
+        public PuzzleDictionary _PuzzleDictionary;
+
+        [Header("Asset References")]
+        public GameObject _BaseView;
+        public TextMeshProUGUI _AttemptsLeftText;
+        public TextMeshProUGUI _LevelsText;
+
+        [Header("Tuning Parameters")]
+        public int _MaxSolveAttempts;
+
 #endregion
 
 #region Private vars
-    private PuzzleBase _activePuzzle;
-    private PuzzleType _activePuzzleType;
-    private string     _activeTriggerKey;
-    private Stack<int> _levelIndexStack;
-    private int        _solveAttemptCount;
-    private int        _startingLevelCount;
-    private Canvas     _canvas;
+
+        private PuzzleBase _activePuzzle;
+        private PuzzleType _activePuzzleType;
+        private string     _activeTriggerKey;
+        private Stack<int> _levelIndexStack;
+        private int        _solveAttemptCount;
+        private int        _startingLevelCount;
+        private Canvas     _canvas;
+
 #endregion
 
 #region Unity API
-    private void Awake()
-    {
-        _canvas = GetComponentInParent<Canvas>();
-        
-        _GameEventDispatcher.AddListener(GameEventType.ShowPuzzleWindow, OnShowPuzzleWindow);
-        _GameEventDispatcher.AddListener(GameEventType.HidePuzzleWindow, OnHidePuzzleWindow);
 
-        OnHidePuzzleWindow(null);
-    }
+        private void Awake()
+        {
+            _canvas = GetComponentInParent<Canvas>();
 
-    private void OnDestroy()
-    {
-        _GameEventDispatcher.RemoveListener(GameEventType.ShowPuzzleWindow, OnShowPuzzleWindow);
-        _GameEventDispatcher.RemoveListener(GameEventType.HidePuzzleWindow, OnHidePuzzleWindow);
-    }
+            _GameEventDispatcher.AddListener(GameEventType.ShowPuzzleWindow, OnShowPuzzleWindow);
+            _GameEventDispatcher.AddListener(GameEventType.HidePuzzleWindow, OnHidePuzzleWindow);
+
+            OnHidePuzzleWindow(null);
+        }
+
+        private void OnDestroy()
+        {
+            _GameEventDispatcher.RemoveListener(GameEventType.ShowPuzzleWindow, OnShowPuzzleWindow);
+            _GameEventDispatcher.RemoveListener(GameEventType.HidePuzzleWindow, OnHidePuzzleWindow);
+        }
+
 #endregion
 
 #region Private Methods
-    private void OnHidePuzzleWindow(GeneralEvent obj)
-    {
-        _BaseView.SetActive(false);
 
-        DestroyPuzzle(_activePuzzle);
-        
-        _activePuzzle = null;
-
-        _levelIndexStack?.Clear();
-    }
-    
-    private void OnShowPuzzleWindow(GeneralEvent obj)
-    {
-        if (_activePuzzle != null)
+        private void OnHidePuzzleWindow(GeneralEvent obj)
         {
-            Debug.LogWarning("Puzzle already active");
-            return;
-        }
-        
-        _BaseView.SetActive(true);
-        
-        _solveAttemptCount = 0;
+            _BaseView.SetActive(false);
 
-        ShowPuzzleArgs args = (ShowPuzzleArgs)obj.data;
-        _activeTriggerKey = args.TriggerKey;
-        _activePuzzleType = args.PuzzleType;
-        
-        _levelIndexStack    = CreateLevelIndexStack(args);
-        _startingLevelCount = _levelIndexStack.Count;
-        
-        GotoNextLevel();
-        
-        UpdateText();
-    }
+            DestroyPuzzle(_activePuzzle);
 
-    private Stack<int> CreateLevelIndexStack(ShowPuzzleArgs args)
-    {
-        args.SpecificLevelList?.Reverse();
-        var levelList = args.SpecificLevelList ?? _PuzzleDictionary.CreateRandomIndexListOfType(args.PuzzleType, args.RandomLevelCount);
-        
-        return new Stack<int>(levelList);
-    }
+            _activePuzzle = null;
 
-    private PuzzleBase CreatePuzzle(PuzzleType type, int levelIndex)
-    {
-        var newPuzzle = _PuzzleDictionary.CreatePuzzleOfType(type, levelIndex, null);
-        
-        // Error handle a null puzzle
-        if (newPuzzle == null)
-        {
-            Debug.LogError($"Puzzle could not be created for puzzle type:{type}");
-            _GameEventDispatcher.DispatchEvent(GameEventType.HidePuzzleWindow);
-            return null;
+            _levelIndexStack?.Clear();
         }
 
-        newPuzzle.WindowCanvas     =  _canvas;
-        newPuzzle.onPuzzleComplete += OnPuzzleCompleted;
-        newPuzzle.Init();
-        
-        return newPuzzle;
-    }
-
-    private void DestroyPuzzle(PuzzleBase puzzle)
-    {
-        if (puzzle != null)
+        private void OnShowPuzzleWindow(GeneralEvent obj)
         {
-            Destroy(puzzle.gameObject);
-        }
-    }
-    
-    private void OnPuzzleCompleted(bool wasSuccess)
-    {
-        if (wasSuccess)
-        {
-            HandleSuccess();
-            return;
-        }
-        
-        _solveAttemptCount++;
-        
-        UpdateText();
-        
-        if(_solveAttemptCount >= _MaxSolveAttempts)
-        {
-            _GameEventDispatcher.DispatchEvent(GameEventType.GameTrigger, CreateTriggerArgs(TriggerType.TerminalLockout));
-            _GameEventDispatcher.DispatchEvent(GameEventType.HidePuzzleWindow);
-        }
-
-
-        void HandleSuccess()
-        {
-        
-             bool canGotoNextLevel = GotoNextLevel();
-            
-            UpdateText();
-            
-            // We have completed all required puzzle levels
-            if (!canGotoNextLevel)
+            if (_activePuzzle != null)
             {
-                _GameEventDispatcher.DispatchEvent(GameEventType.GameTrigger, CreateTriggerArgs(TriggerType.PuzzleSolved));
+                Debug.LogWarning("Puzzle already active");
+                return;
+            }
+
+            _BaseView.SetActive(true);
+
+            _solveAttemptCount = 0;
+
+            ShowPuzzleArgs args = (ShowPuzzleArgs)obj.data;
+            _activeTriggerKey = args.TriggerKey;
+            _activePuzzleType = args.PuzzleType;
+
+            _levelIndexStack    = CreateLevelIndexStack(args);
+            _startingLevelCount = _levelIndexStack.Count;
+
+            GotoNextLevel();
+
+            UpdateText();
+        }
+
+        private Stack<int> CreateLevelIndexStack(ShowPuzzleArgs args)
+        {
+            args.SpecificLevelList?.Reverse();
+            var levelList = args.SpecificLevelList ?? _PuzzleDictionary.CreateRandomIndexListOfType(args.PuzzleType, args.RandomLevelCount);
+
+            return new Stack<int>(levelList);
+        }
+
+        private PuzzleBase CreatePuzzle(PuzzleType type, int levelIndex)
+        {
+            var newPuzzle = _PuzzleDictionary.CreatePuzzleOfType(type, levelIndex, null);
+
+            // Error handle a null puzzle
+            if (newPuzzle == null)
+            {
+                Debug.LogError($"Puzzle could not be created for puzzle type:{type}");
                 _GameEventDispatcher.DispatchEvent(GameEventType.HidePuzzleWindow);
+                return null;
+            }
+
+            newPuzzle.WindowCanvas     =  _canvas;
+            newPuzzle.onPuzzleComplete += OnPuzzleCompleted;
+            newPuzzle.Init();
+
+            return newPuzzle;
+        }
+
+        private void DestroyPuzzle(PuzzleBase puzzle)
+        {
+            if (puzzle != null)
+            {
+                Destroy(puzzle.gameObject);
             }
         }
 
-        GameTriggerArgs CreateTriggerArgs(TriggerType type)
+        private void OnPuzzleCompleted(bool wasSuccess)
         {
-            return new GameTriggerArgs
+            if (wasSuccess)
             {
-                TriggerType = type,
-                TriggerKey  = _activeTriggerKey,
-            };
-        }
-    }
-    
-    private bool GotoNextLevel()
-    {
-        if (!_levelIndexStack.TryPop(out int nextLevelIndex))
-        {
-            return false;
-        }
-        
-        DestroyPuzzle(_activePuzzle);
-        _activePuzzle = CreatePuzzle(_activePuzzleType, nextLevelIndex);
-        
-        return true;
-    }
+                HandleSuccess();
+                return;
+            }
 
-    private void UpdateText()
-    {
-        _AttemptsLeftText.text = $"Attempts Left: {_MaxSolveAttempts - _solveAttemptCount}";
-        _LevelsText.text       = $"Levels: {_startingLevelCount - _levelIndexStack.Count}/{_startingLevelCount}";
-    }
+            _solveAttemptCount++;
+
+            UpdateText();
+
+            if (_solveAttemptCount >= _MaxSolveAttempts)
+            {
+                _GameEventDispatcher.DispatchEvent(GameEventType.GameTrigger, CreateTriggerArgs(TriggerType.TerminalLockout));
+                _GameEventDispatcher.DispatchEvent(GameEventType.HidePuzzleWindow);
+            }
+
+
+            void HandleSuccess()
+            {
+                bool canGotoNextLevel = GotoNextLevel();
+
+                UpdateText();
+
+                // We have completed all required puzzle levels
+                if (!canGotoNextLevel)
+                {
+                    _GameEventDispatcher.DispatchEvent(GameEventType.GameTrigger, CreateTriggerArgs(TriggerType.PuzzleSolved));
+                    _GameEventDispatcher.DispatchEvent(GameEventType.HidePuzzleWindow);
+                }
+            }
+
+            GameTriggerArgs CreateTriggerArgs(TriggerType type)
+            {
+                return new GameTriggerArgs
+                {
+                    TriggerType = type,
+                    TriggerKey  = _activeTriggerKey,
+                };
+            }
+        }
+
+        private bool GotoNextLevel()
+        {
+            if (!_levelIndexStack.TryPop(out int nextLevelIndex))
+            {
+                return false;
+            }
+
+            DestroyPuzzle(_activePuzzle);
+            _activePuzzle = CreatePuzzle(_activePuzzleType, nextLevelIndex);
+
+            return true;
+        }
+
+        private void UpdateText()
+        {
+            _AttemptsLeftText.text = $"Attempts Left: {_MaxSolveAttempts - _solveAttemptCount}";
+            _LevelsText.text       = $"Levels: {_startingLevelCount - _levelIndexStack.Count}/{_startingLevelCount}";
+        }
+
 #endregion
+    }
 }
